Move stop logic back to sensible location
I had moved the process monitoring into start as I had the incorrect belief that it was necessary to prevent zombie processes. I have since moved it back so that the logic is clearer. I have also re-instated the use of a single state exposure channel. This channel is used only to inform external code of ready states but not of changes of state from Running to Stopping. It only informs of state changes such as Running to Stopped or Stopped to Running.