Introduce Bootstrapper
Bootstrapper makes it easier to implement an Instance without re-writting the core aync mechanics which would be a shame as they are all tested. So now you can create an instance using a boostrapper and a reference string and that will then use the bootstrapper to create the actual process but continue using the tested logic for starting, stopping and messaging.